move ip6 commands to inet256ipv6 package
